Fixed parsing bug

+++ SBG/t/NetworkIO.t
@@ -4,7 +4,7 @@ use Test::More 'no_plan';
 
 use SBG::NetworkIO;
 
-my $file = "$installdir/t/ex_templates_descriptors.csv";
+my $file = shift || "$installdir/t/ex_templates_descriptors.csv";
 
 my $io = new SBG::NetworkIO(-file=>$file);
 my $net = $io->read;
@@ -13,7 +13,7 @@ my $net = $io->read;
 my $graphout = "$installdir/t/graph.png";
 SBG::NetworkIO::graphviz($net, $graphout,-edge_color=>'grey');
 ok(-r $graphout, "GraphViz PNG creation");
-unlink $graphout;
+unlink $graphout unless shift;
 
 __END__
